requirements.txt
pyserial
cobs
pyserial>=3.5